[x264] fix compilation errors on Linux

+++ branches/avidemux_2.5_branch_gruntster/avidemux/plugins/ADM_videoEncoder/ADM_vidEnc_x264/options.cpp
@@ -748,7 +748,7 @@ void x264Options::setQuantiser(unsigned int quantiser)
 
 unsigned int x264Options::getVbrQuantiser(void)
 {
-	return _param.rc.f_rf_constant;
+	return (unsigned int)_param.rc.f_rf_constant;
 }
 
 void x264Options::setVbrQuantiser(unsigned vbrQuantiser)
